small fixes

+++ clip_generator/editter/audio_info.py
@@ -60,12 +60,12 @@ def write_infos_edit():
         print(str(info[0][1]['pad']) + " - " + str(info[0][1]['pad_post']))
 
 
-def write_infos_trim(from_second: int, to_second: int):
+def write_infos_trim(from_second: float, to_second: float):
     print(str(from_second) + " - " + str(to_second))
     appendJSON({'trim': [from_second, to_second]})
 
 
-def write_correlation(start: int, end: int):
+def write_correlation(start: float, end: float):
     print({'correlation': {'trim': [start, end]}})
     appendJSON({'correlation': {'trim': [start, end]}})
